% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advocates.R
\name{advocates}
\alias{advocates}
\alias{advocates.OyezCase}
\alias{advocates.OyezCaseList}
\title{Get advocates for an \code{Oyez} object}
\usage{
advocates(x, ...)

\method{advocates}{OyezCase}(x, ...)

\method{advocates}{OyezCaseList}(x, ...)
}
\arguments{
\item{x}{the Oyez object}

\item{...}{other arguments passed on to methods}
}
\value{
an object of class OyezAdvocateList
}
\description{
Get advocates for an \code{Oyez} object
}
