% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OyezTranscript_methods.R
\name{as.data.frame.OyezTranscript}
\alias{as.data.frame.OyezTranscript}
\title{Converts an OyezTranscript object into a data.frame}
\usage{
\method{as.data.frame}{OyezTranscript}(x, ...)
}
\arguments{
\item{x}{an OyezTranscript object}

\item{...}{other arguments passed on to methods}
}
\value{
a data.frame object
}
\description{
Converts an OyezTranscript object into a data.frame
}
\examples{
\dontrun{
a <- oyez_case(term = 2014, docket_number='13-1175')
t <- transcript(a)
as.data.frame(t)
}
}
