% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyez_case.R
\name{get_oyez_case_}
\alias{get_oyez_case_}
\title{Splitting this internal function so it can work with other methods}
\usage{
get_oyez_case_(case_url)
}
\description{
Splitting this internal function so it can work with other methods
}
\keyword{internal}
