% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyez_case.R
\name{oyez_case}
\alias{oyez_case}
\title{Get a single Oyez Case}
\usage{
oyez_case(term, docket_number)
}
\arguments{
\item{term}{the term}

\item{docket_number}{the docket number}
}
\value{

}
\description{
Get a single Oyez Case
}
\examples{
\dontrun{oyez_case(term = 2014, docket_number='13-1175')}

}
