% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyez_case.R
\name{oyez_cases}
\alias{oyez_cases}
\alias{oyez_cases.data.frame}
\alias{oyez_cases.list}
\title{Get multiple oyez cases}
\usage{
oyez_cases(x, ...)

\method{oyez_cases}{data.frame}(x, ...)

\method{oyez_cases}{list}(x, ...)
}
\arguments{
\item{x}{a data.frame or list}

\item{...}{}
}
\value{
an OyezCaseList object
}
\description{
Get multiple oyez cases
}
