% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript.R
\name{transcript}
\alias{transcript}
\alias{transcript.OyezCase}
\alias{transcript.OyezCaseList}
\title{Gets transcripts for OyezCase objects or OyezCaseLists}
\usage{
transcript(x, ...)

\method{transcript}{OyezCase}(x, ...)

\method{transcript}{OyezCaseList}(x, ...)
}
\arguments{
\item{x}{an OyezCase or OyezCaseList object}

\item{...}{other arguments passed on to methods}
}
\value{
an OyezTranscript or OyezTranscriptList object
}
\description{
Gets transcripts for OyezCase objects or OyezCaseLists
}
\examples{
\dontrun{
a <- oyez_case(term = 2014, docket_number='13-1175')
transcript(a)
}
}
