% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{cases}
\alias{cases}
\title{Get cases for an \code{oyez} object}
\usage{
cases(x, ...)
}
\arguments{
\item{x}{the oyez object}

\item{...}{other arguments}
}
\description{
Get cases for an \code{oyez} object
}
