% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{cases}
\alias{cases}
\alias{cases.OyezAdvocate}
\alias{cases.OyezAdvocateList}
\title{Get cases for an \code{Oyez} object}
\usage{
cases(x, ...)

\method{cases}{OyezAdvocate}(x, ...)

\method{cases}{OyezAdvocateList}(x, ...)
}
\arguments{
\item{x}{the Oyez object}

\item{...}{other arguments passed on to methods}
}
\value{
an object of class OyezCaseList
}
\description{
Get cases for an \code{Oyez} object
}
