% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyez_case.R
\name{oyez_cases}
\alias{oyez_cases}
\alias{oyez_cases.data.frame}
\alias{oyez_cases.list}
\title{Get multiple oyez cases}
\usage{
oyez_cases(x, ...)

\method{oyez_cases}{data.frame}(x, ...)

\method{oyez_cases}{list}(x, ...)
}
\arguments{
\item{x}{a data.frame or list}

\item{...}{other arguments passed on to methods}
}
\value{
an OyezCaseList object
}
\description{
Gets multiple oyez cases for analysis. The input, x, can be either a data.frame with columns
\code{term} and \code{docket_number}, or a list of lists, with each sublist containing fields
named \code{term} and \code{docket_number}
}
\examples{
\dontrun{
 df <- data.frame(
     term = c("2014", "2014", "2000"),
     docket_number = c("13-1175", "13-1499", "00-949")
 )
 oyez_cases(df)

 lst <- list(
       list(
           term = "2014",
           docket_number = "13-1175"
       ),
       list(
           term = "2014",
           docket_number = "13-1499"
       ),
       list(
           term = "2000",
           docket_number = "00-949"
       )
   )
 oyez_cases(lst)
}

}
\keyword{internal}
